pyproject.toml
[tool.poetry]
name = "slackblocks"
version = "0.9.6"
description = "Python wrapper for the Slack Blocks API"
authors = [
    "Nicholas Lambourne <dev@ndl.im>",
]
maintainers = [
    "Nicholas Lambourne <dev@ndl.im>",
]
homepage = "https://github.com/nicklambourne/slackblocks"
repository = "https://github.com/nicklambourne/slackblocks"
license = "BSD-3-Clause"
readme = "README.md"
keywords = [
    "slackblocks", 
    "slack", 
    "messaging", 
    "message generation", 
    "slack blocks", 
    "blocks",
]
classifiers=[
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Typing :: Typed",
    "Topic :: Communications :: Chat"
]
exclude = ["test/**", "docs/**"]

[tool.poetry.dependencies]
python = ">=3.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
twine = "^4.0.2"
wheel = "^0.40.0"
slackclient = "^2.9.4" 
black = "^23.1.0"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
exclude = [".venv", "./build", "./dist", ".eggs", ".git"]
max-line-length = 100
per-file-ignores = [
    "slackblocks/__init__.py:F401",
]
